Add a dark mode switch web component
This commit adds a web component that renders a switch designed to
toggle between the light and dark theme of a website. It uses a moon to
symbolize the dark theme and a sun to symbolize the light theme.

These symbols are embedded SVG images that have been designed from
scratch, i.e. the same copyright applies to those images as for the
entire web component.