Add a discrete slider web component
This commit adds a web component that implements a slider element with
functionality that is comparable to the standard HTML "input" element of
type "range".

However, the standard HTML range input element suffers from a number of
limitations that justified implementing this web component. First, it is
unnecessarily complicated to style the standard HTML range input element
to make it look the same in all web browsers (see [1] and [2]).

Second, the standard HTML range input element is designed to be used in
a horizontal manner from left to right. Turning it into a vertical
slider can be done but is again web browser-specific (see [3]). In order
to invert its minimum and maximum the element needs to be transformed
accordingly (see [4]).

Lastly, the standard HTML range input element does not support visual
indications of the values that it can be set to. Such indications are an
important aspect of this web component and it is where it derives its
name from.

[1] https://css-tricks.com/sliding-nightmare-understanding-range-input/
[2] https://css-tricks.com/styling-cross-browser-compatible-range-inputs-css/
[3] https://stackoverflow.com/q/15935837
[4] http://twiggle-web-design.com/tutorials/Custom-Vertical-Input-Range-CSS3.html