Make the slider key bindings orientation-dependent
Before this commit, the arrow up and arrow right keys were always used
to increment the slider's value while the arrow down and arrow left keys
were always used to decrement the slider's value.

The approach proofed to be confusing for all slider orientations except
left-to-right. To address that confusion, this commit changes the arrow
key bindings depending on the slider's orientation.

In addition, the Home and End keys are now used as well. Their behavior
mimics the behavior of the standard HTML "input" element of type
"range", i.e. Home jumps to the minimum and End to the maximum value.

To prevent the default actions to be executed if the arrow buttons are
used, the "Event.preventDefault" method is now called.