Define the properties using computed property names
Computed property names is a feature of ECMAScript 2015 (ES6) that
allows putting an expression in square brackets which JavaScript then
uses as the property name (see [1]).

This feature is used by this commit to define the property names using
the constants that are already used to define the attributes. This way,
both the property names and the attribute names can be kept in sync
without requiring any additional efforts.

[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Object_initializer#computed_property_names