Change the thumb position in the CSS rule of the host
Before this commit, the value of the CSS custom property that defines
the position of the thumb was set on the track element. Because the
thumb is a child element of the track that custom property was simply
inherited.

However, all other CSS custom properties are defined in the ":host" rule
of the CSS style sheet of the web component. This commit changes the
approach so that the CSS custom property storing the thumb position is
now defined in the ":host" rule as well.

This is achieved by directly modifying the style sheet of the shadow DOM
via corresponding APIs.