Minimize the amount of reflows caused by the slider
According to Google's documentation, "reflow is the name of the web
browser process for re-calculating the positions and geometries of
elements in the document, for the purpose of re-rendering part or all of
the document" (see [1]).

Minimizing the amount of reflows reduces the amount of time that the web
browser is blocked which means it is critical for overall website
performance.

Reflows are triggered when adding to or removing elements from the
document which are operations that the web component requires to adjust
the amount of dots that are visible on the slider.

The current approach used by the web component is to add every dot
element to the document individually which causes a reflow for each dot.
The same approach is taken if dot elements need to be removed.

To reduce the number of reflows, that approach is changed so that all
dots are first added to a DocumentFragment (see [2]). A DocumentFragment
is not part of the active document which means no reflows are triggered
when adding to or removing elements from it.

In a second step, all dot elements in the DocumentFragment are moved to
the document in a single operation which only triggers a single reflow
(see [2]). At the same time, dot elements that might already be part of
the document are replaced.

[1] https://developers.google.com/speed/docs/insights/browser-reflow
[2] https://developer.mozilla.org/en-US/docs/Web/API/DocumentFragment