Expose the orientation enum via static property
Before this commit, the orientation enum was exposed to the end user of
the module as a second export. This commit removes that export and
instead exposes the enum as a static property of the slider class
because this approach offers a number of advantages.

First, the interface of the module gets simpler because there is only
one default export. Second, it is more obvious what the context of the
orientation enum is. Lastly, the orientation enum can be renamed when
minifying the file without having an impact on the module interface.