Explicitly provide the radix when parsing into integer
Google's Closure Compiler emits the following warning when it encounters
a call of the parseInt function without the second argument:

"WARNING - [JSC_WRONG_ARGUMENT_COUNT] Function parseInt: called with 1
argument(s). Function requires at least 2 argument(s) and no more than 2
argument(s)."

According to [1], the second argument is the radix. It is optional but
does not always default to 10. This commit makes it explicit that a
radix of 10 is to be used.

[1] https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/parseInt