Define the minimum size on the slider itself
Before this commit, a minimum size was only defined for the track of the
slider but not for the slider itself. Because of this, the width and
height of the slider could be defined to be smaller than the size
required by the track.

To prevent this, this commit adds both the min-width and min-height CSS
properties to the slider itself and adjusts their values if the
orientation of the slider is changed.

Note that the min-width and min-height CSS properties of the track must
still be kept because a child of a parent with min-height set does not
inherit the height of that parent (see [1]).

In addition, this commit changes the minimum size of the slider to the
smallest possible size that still renders correctly if only two dots are
drawn. This is done to support sliders with a length smaller than the
previously given 150px.

To make the default size of the slider more meaningful than the minimum
size, an initial size is introduced. If no size is specifically defined
for the slider that initial size will be used.

[1] https://stackoverflow.com/q/8468066