Base the progress calculation on the host's dimensions
This commit adjusts the calculation of the progress so that the
dimensions of the web component itself are used instead of relying on
the dimensions provided by the "track" div element. The result of the
calculation does not change, only how it is calculated changes.

Not relying on a child component, whose dimensions directly depend on
the web component itself, seems to make the calculation more straight-
forward which is the reason for this change.

The calculation operates with a number of properties that are explained
by the below image:

                      event.offsetX
          |<--------------------------------->|
          |                                   |
          |    event.clientX                  |
----------+---------------------------------->|
          |                                   |
          |                                   |
  +-------+--------------------margin---------+--------------------+
  |       |                                   |                    |
  |   +---+--------------------border---------+---------------+----+-------------------------------
  |   |   |                                   |               |    |                             ^
  |   |   +--------------------padding--------+-----------+---+----+---------------              |
  |   |   |                                   |           |   |    |             ^               |
  |   |   |   +----------------content--------+-------+---+---+----+-----        |               |
  |   |   |   | +--------+--------+--------+--x-----+ |   |   |    |   | height  | clientHeight  | offsetHeight
  |   |   |   +-+-----------------------------------+-+---+---+----+-----        |               |
  |   |   |   | |                                   | |   |   |    |             v               |
  |   |   +---+-+-----------------------------------+-+---+---+----+---------------              |
  |   |   |   | |                                   | |   |   |    |                             v
  |   +---+---+-+-----------------------------------+-+---+---+----+-------------------------------
  |   |   |   | |                                   | |   |   |    |
  +---+---+---+-+-----------------------------------+-+---+---+----+
      |   |   | |                                   | |   |   |
      |   |   | |            trackLength            | |   |   |
      |   |   | |<--------------------------------->| |   |   |
      |   |   |                                       |   |   |
      |   |   |                 width                 |   |   |
      |   |   |<------------------------------------->|   |   |
      |   |                                               |   |
      |   |                  clientWidth                  |   |
      |   |<--------------------------------------------->|   |
      |                                                       |
      |                      offsetWidth                      |
      |<----------------------------------------------------->|

First, the web component's width and height are used which are accessed
using the computed style of the web component. In comparison, both
"clientWidth" and "clientHeight" both include padding but exclude
borders and margins (see [1]) while the previously used "offsetWidth"
and "offsetHeight" include both padding and borders but not the margins
(see [2]).

Second, the coordinates of the pointer event are stored in "offsetX" and
"offsetY" properties. Those coordinates are measured from the edge that
marks the beginning of the padding, i.e. they are measured from the same
baseline as the clientWidth (see [3] and [4]).

Using these dimensions together with the padding dimensions enables the
calculation of the progress without relying on child elements. This
modified calculation allows setting a padding and border on the slider
itself which is why those CSS properties are removed.

[1] https://developer.mozilla.org/en-US/docs/Web/API/Element/clientWidth
[2] https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/offsetWidth
[3] https://developer.mozilla.org/en-US/docs/Web/API/MouseEvent/offsetX
[4] https://developer.mozilla.org/en-US/docs/Web/API/MouseEvent/offsetY