Replace usages of the deprecated srcElement property
JavaScript's event interface supports a number of different properties
that return HTML elements that are involved in the event.

First, "event.target" returns the HTML element that actually caused the
event to be fired. For example, if a button has an event listener
attached and that button gets clicked the "target" property of the event
that is passed as argument to the event listener is a reference to the
button.

However, if the button contains a nested SVG image and a path of that
SVG image is clicked "event.target" will return a reference to that
particular path object, not to the button that contains the SVG image.

Second, "event.currentTarget" returns the HTML element to which the
event listener is attached. In the above example, "event.currentTarget"
will always return a reference to the button, independent of whether the
button itself or a part of the SVG image is clicked. That is the case
because the event listener is attached to the button.

Third, "event.srcElement" is deprecated according to [1]. At the time of
this commit that property still works but it is no longer recommended to
be used. Apparently, it is an alias for the "event.target" property.

To address the deprecation, this commit replaces all usages of
"event.srcElement" with "event.currentTarget". The "currentTarget"
property is used instead of the "target" property because in all cases
that are replaced it is important to get the element to which the event
listener is attached.

[1] https://developer.mozilla.org/en-US/docs/Web/API/Event/srcElement