Prevent the pointer down event from bubbling up
Subsequent pointer down events on the dark mode switch might be
interpreted by the web browser to be double clicks if they are fast
enough. Depending on the web page that the switch is located on, those
double clicks can result in text being selected.

To prevent text from being selected, pointer click events that occur on
the dark mode switch are now swallowed, i.e. prevented from bubbling up
the hierarchy of elements.

Another option would be to prevent the default action associated with an
event by calling "Element.preventDefault()" in the event handler.
However, letting the event bubble up and telling elements higher in the
hierarchy not to execute their default handler seems to make less sense
than to just prevent the event from bubbling up in the first place (see
[1]).

[1] https://stackoverflow.com/a/68442418