Make the Web Components programmatically creatable
Before this commit, creating instances of the web components via
JavaScript resulted in type errors because the style sheet that is
retrieved from the shadow root is not accessible in the constructor of
the web components.

This issue only occurs if instances are created using JavaScript. Using
the web components directly in HTML works as expected. It is save to
access the style sheet in the "connectedCallback". However, there is no
real benefit to maintaining properties storing references to style sheet
parts anyway.

Instead, the style sheet is now only accessed if it needs to be
modified. This prevents the type errors from occurring. However, the
discrete slider modifies properties of the style sheet as part of the
attributeChangedCallback.

This callback is executed before the connectedCallback at a point in
time when the style sheet isn't available yet. This means that the
corresponding operations cannot be executed at that point in time
without encountering type errors. Hence, an if-statement is added that
prevents the execution of those operations.

The corresponding operations are then executed again in the
connectedCallback which is save because the style sheet does exist at
that point in time. This ensures that the web components can now be
created and added to the DOM using JavaScript without running into type
errors.

To verify this, the HTML files testing the web components have been
extended to include JavaScript code that creates and adds instances
programmatically.