Replace the deprecated CSSStyleSheet rules property
The Cascading Style Sheet Object Model specification mandates that the
"rules" attribute of the "CSSStyleSheet" interface is deprecated (see
[1]). It states that "the rules attribute must follow the same steps as
cssRules, and return the same object cssRules would return".

Therefore, the "cssRules" attribute is a replacement for the "rules"
attribute and this commit replaces the use of the "rules" property with
the "cssRules" property in the JavaScript code.

[1]: https://www.w3.org/TR/2021/WD-cssom-1-20210826/#the-cssstylesheet-interface