Use ECMAScript 2022 private class fields and methods
Before this commit, an underscore character (_) was used to mark class
members that are not supposed to be accessed from outside of the class.
However, this approach is just a convention and cannot actually prevent
members from being accessed from outside of the class.

ECMAScript 2022 introduced private class members (see [1]) which means
that JavaScript itself will now enforce the privacy of members. This
makes the previously used convention obsolete and thus the underscore
character is replaced with a hash character (#) which turns the member
into a private member.

According to [2], this ECMAScript 2022 feature is implemented in all
major web browsers at the time of this commit which is why this commit
adjusts the JavaScript code to make use of this feature.

[1]: https://262.ecma-international.org/13.0/#prod-PrivateIdentifier
[2]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Classes/Private_class_fields