Add a script to automatically minify the web components
Before this commit, the minified versions of the JavaScript files were
created manually. This was done because the HTML and CSS is embedded in
the JavaScript file as a template literal and as such is exempt by
JavaScript minifiers.

This commit adds a JavaScript file that is specifically designed to
minify web components that contain a template literal with HTML and CSS.
It uses a Regular Expression to extract the template literal, minify it
separately from the JavaScript code and then outputs the reassembled web
component.

To minify the JavaScript, the script relies on Terser (see [1]) which
must be installed using npm before running the script. This can be done
by executing the command "npm install".

The commit also adds the minified web components that were created by
invoking that JavaScript file.

[1]: https://github.com/terser/terser